fix(type-promotion): implement NEP 50 for unsigned array + signed scalar
NumPy 2.0 introduced NEP 50 which changed type promotion rules: when an
unsigned integer array operates with a signed integer scalar of the same
or lower "kind", the array dtype now wins (no type widening).

This updates _typemap_arr_scalar for 12 entries:
- uint8 + int16/int32/int64 → uint8 (was int16/int32/int64)
- uint16 + int16/int32/int64 → uint16 (was int32/int32/int64)
- uint32 + int16/int32/int64 → uint32 (was int64/int64/int64)
- uint64 + int16/int32/int64 → uint64 (was float64/float64/float64)

Breaking change for users relying on NumPy 1.x widening behavior.
This aligns with NumPy 2.x and the project's NumPy alignment goal.

Closes #529

See: https://numpy.org/neps/nep-0050-scalar-promotion.html

+++ src/NumSharp.Core/Logic/np.find_common_type.cs
@@ -259,11 +259,11 @@ namespace NumSharp
             typemap_arr_scalar.Add((np.uint8, np.@bool), np.uint8);
             typemap_arr_scalar.Add((np.uint8, np.uint8), np.uint8);
             typemap_arr_scalar.Add((np.uint8, np.@char), np.uint8);
-            typemap_arr_scalar.Add((np.uint8, np.int16), np.int16);
+            typemap_arr_scalar.Add((np.uint8, np.int16), np.uint8);  // NEP 50: array dtype wins
             typemap_arr_scalar.Add((np.uint8, np.uint16), np.uint8);
-            typemap_arr_scalar.Add((np.uint8, np.int32), np.int32);
+            typemap_arr_scalar.Add((np.uint8, np.int32), np.uint8);  // NEP 50: array dtype wins
             typemap_arr_scalar.Add((np.uint8, np.uint32), np.uint8);
-            typemap_arr_scalar.Add((np.uint8, np.int64), np.int64);
+            typemap_arr_scalar.Add((np.uint8, np.int64), np.uint8);  // NEP 50: array dtype wins
             typemap_arr_scalar.Add((np.uint8, np.uint64), np.uint8);
             typemap_arr_scalar.Add((np.uint8, np.float32), np.float32);
             typemap_arr_scalar.Add((np.uint8, np.float64), np.float64);
@@ -298,11 +298,11 @@ namespace NumSharp
             typemap_arr_scalar.Add((np.uint16, np.@bool), np.uint16);
             typemap_arr_scalar.Add((np.uint16, np.uint8), np.uint16);
             typemap_arr_scalar.Add((np.uint16, np.@char), np.uint16);
-            typemap_arr_scalar.Add((np.uint16, np.int16), np.int32);
+            typemap_arr_scalar.Add((np.uint16, np.int16), np.uint16);  // NEP 50: array dtype wins
             typemap_arr_scalar.Add((np.uint16, np.uint16), np.uint16);
-            typemap_arr_scalar.Add((np.uint16, np.int32), np.int32);
+            typemap_arr_scalar.Add((np.uint16, np.int32), np.uint16);  // NEP 50: array dtype wins
             typemap_arr_scalar.Add((np.uint16, np.uint32), np.uint16);
-            typemap_arr_scalar.Add((np.uint16, np.int64), np.int64);
+            typemap_arr_scalar.Add((np.uint16, np.int64), np.uint16);  // NEP 50: array dtype wins
             typemap_arr_scalar.Add((np.uint16, np.uint64), np.uint16);
             typemap_arr_scalar.Add((np.uint16, np.float32), np.float32);
             typemap_arr_scalar.Add((np.uint16, np.float64), np.float64);
@@ -324,11 +324,11 @@ namespace NumSharp
             typemap_arr_scalar.Add((np.uint32, np.@bool), np.uint32);
             typemap_arr_scalar.Add((np.uint32, np.uint8), np.uint32);
             typemap_arr_scalar.Add((np.uint32, np.@char), np.uint32);
-            typemap_arr_scalar.Add((np.uint32, np.int16), np.int64);
+            typemap_arr_scalar.Add((np.uint32, np.int16), np.uint32);  // NEP 50: array dtype wins
             typemap_arr_scalar.Add((np.uint32, np.uint16), np.uint32);
-            typemap_arr_scalar.Add((np.uint32, np.int32), np.int64);
+            typemap_arr_scalar.Add((np.uint32, np.int32), np.uint32);  // NEP 50: array dtype wins
             typemap_arr_scalar.Add((np.uint32, np.uint32), np.uint32);
-            typemap_arr_scalar.Add((np.uint32, np.int64), np.int64);
+            typemap_arr_scalar.Add((np.uint32, np.int64), np.uint32);  // NEP 50: array dtype wins
             typemap_arr_scalar.Add((np.uint32, np.uint64), np.uint32);
             typemap_arr_scalar.Add((np.uint32, np.float32), np.float64);
             typemap_arr_scalar.Add((np.uint32, np.float64), np.float64);
@@ -350,11 +350,11 @@ namespace NumSharp
             typemap_arr_scalar.Add((np.uint64, np.@bool), np.uint64);
             typemap_arr_scalar.Add((np.uint64, np.uint8), np.uint64);
             typemap_arr_scalar.Add((np.uint64, np.@char), np.uint64);
-            typemap_arr_scalar.Add((np.uint64, np.int16), np.float64);
+            typemap_arr_scalar.Add((np.uint64, np.int16), np.uint64);  // NEP 50: array dtype wins
             typemap_arr_scalar.Add((np.uint64, np.uint16), np.uint64);
-            typemap_arr_scalar.Add((np.uint64, np.int32), np.float64);
+            typemap_arr_scalar.Add((np.uint64, np.int32), np.uint64);  // NEP 50: array dtype wins
             typemap_arr_scalar.Add((np.uint64, np.uint32), np.uint64);
-            typemap_arr_scalar.Add((np.uint64, np.int64), np.float64);
+            typemap_arr_scalar.Add((np.uint64, np.int64), np.uint64);  // NEP 50: array dtype wins
             typemap_arr_scalar.Add((np.uint64, np.uint64), np.uint64);
             typemap_arr_scalar.Add((np.uint64, np.float32), np.float64);
             typemap_arr_scalar.Add((np.uint64, np.float64), np.float64);
